containers/odata-proxy/requirements.txt
Flask==2.0.0
PyJWT
cryptography
requests
PyOpenSSL
gunicorn==20.0.4